+++ CHANGES.rst
@@ -4,7 +4,10 @@ Changelog
 1.7.0+enfold9 (unreleased)
 --------------------------
 
-- Nothing changed yet.
+- FD#1756: If the object referenced from a related items field is an image,
+  get the preview scale instead of the full sized image. This can optionally
+  be controlled using an 'image.scale' request header with the scale name
+  [frapell]
 
 
 1.7.0+enfold8 (2016-06-10)